Move test-only fixtures (NullSandbox, DummyWorkflow) to tests/fakes.py
Both lived under src/cog/ but were only imported by tests. Moving
them next to the rest of the test fakes keeps src/ aligned with
production code and removes two whole modules from import-graph
analysis.
