+++ tests/test_runners/test_agent_runner_run_default.py
@@ -2,7 +2,7 @@
 
 from cog.core.runner import RunResult
 from cog.runners.claude_cli import ClaudeCliRunner
-from cog.runners.sandbox import NullSandbox
+from tests.fakes import NullSandbox
 from tests.test_runners.helpers import fixture_proc, patch_exec
