+++ tests/test_runners/test_claude_cli_result.py
@@ -4,7 +4,7 @@ import pytest
 
 from cog.core.runner import ResultEvent
 from cog.runners.claude_cli import ClaudeCliRunner
-from cog.runners.sandbox import NullSandbox
+from tests.fakes import NullSandbox
 from tests.test_runners.helpers import fixture_proc, patch_exec
